RUB-243: extract sync_pv.go — PV shadow helpers from sync.go (813→714 lines)
